+++ packages/client/src/lib/info.json
@@ -1,1 +1,1 @@
-{ "version": "1.0.6" }
+{"version":"1.0.6"}
